+++ application/language/russian/datatable_lang.php
@@ -1,36 +1,40 @@
-<?php if (!defined('BASEPATH')) exit('No direct script access allowed');
-/*
-	Copyright (c) 2011 Lonnie Ezell
+<?php defined('BASEPATH') || exit('No direct script access allowed');
+/**
+ * Bonfire
+ *
+ * An open source project to allow developers to jumpstart their development of
+ * CodeIgniter applications.
+ *
+ * @package   Bonfire
+ * @author    Bonfire Dev Team
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
+ * @link      http://cibonfire.com
+ * @since     Version 1.0
+ */
 
-	Permission is hereby granted, free of charge, to any person obtaining a copy
-	of this software and associated documentation files (the "Software"), to deal
-	in the Software without restriction, including without limitation the rights
-	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-	copies of the Software, and to permit persons to whom the Software is
-	furnished to do so, subject to the following conditions:
+/**
+ * Application/datatable language file (Russian).
+ *
+ * Localization strings used by Bonfire.
+ *
+ * @package Bonfire\Application\Language\Russian
+ * @author  Translator < https://github.com/modgahead >
+ * @link    http://cibonfire.com/docs/guides
+ */
 
-	The above copyright notice and this permission notice shall be included in
-	all copies or substantial portions of the Software.
+$lang['sProcessing']   = 'В процессе...';
+$lang['sLengthMenu']   = 'Показывать записи _MENU_';
+$lang['sZeroRecords']  = 'Записей не найдено';
+$lang['sInfo']         = 'Показывать с _START_ по _END_ из _TOTAL_ записей';
+$lang['sInfoEmpty']    = 'Показывать с 0 по 0 из 0 записей';
+$lang['sInfoFiltered'] = '(отфильтровано из _MAX_ записей)';
+//$lang['sInfoPostFix']  = '';
+$lang['sSearch']       = 'Поиск:';
+//$lang['sUrl']		   = '';
+$lang['sFirst']        = 'Первая';
+$lang['sPrevious']     = 'Предыдущая';
+$lang['sNext']         = 'Следующая';
+$lang['sLast']         = 'Последняя';
 
-	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-	THE SOFTWARE.
-*/
-
-$lang['sProcessing']	= 'В процессе...';
-$lang['sLengthMenu']	= 'Показывать записи _MENU_';
-$lang['sZeroRecords']	= 'Записей не найдено';
-$lang['sInfo']			= 'Показывать с _START_ по _END_ из _TOTAL_ записей';
-$lang['sInfoEmpty']		= 'Показывать с 0 по 0 из 0 записей';
-$lang['sInfoFiltered']	= '(отфильтровано из _MAX_ записей)';
-$lang['sInfoPostFix']	= '';
-$lang['sSearch']		= 'Поиск:';
-$lang['sUrl']			= '';
-$lang['sFirst']			= 'Первая';
-$lang['sPrevious']		= 'Предыдущая';
-$lang['sNext']			= 'Следующая';
-$lang['sLast']			= 'Последняя';
+/* End of file ./application/language/russian/datatable_lang.php */