+++ bonfire/modules/activities/language/russian/activities_lang.php
@@ -1,0 +1,76 @@
+<?php defined('BASEPATH') || exit('No direct script access allowed');
+/**
+ * Bonfire
+ *
+ * An open source project to allow developers to jumpstart their development of
+ * CodeIgniter applications
+ *
+ * @package   Bonfire
+ * @author    Bonfire Dev Team
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
+ * @link      http://cibonfire.com
+ * @since     Version 1.0
+ */
+
+/**
+ * Activities language file (Russian)
+ *
+ * Localization strings used by Bonfire
+ *
+ * @package    Bonfire\Modules\Activities\Language\Russian
+ * @author     Translator < https://github.com/cjmaxik >
+ * @link       http://cibonfire.com/docs/guides
+ */
+
+$lang['activity_home']                  = 'Домой';
+$lang['activity_title']                 = 'Логи действий';
+$lang['activity_own']                   = 'Вы';
+$lang['activity_user']                  = 'Пользователь';
+$lang['activity_users']                 = 'Пользователи';
+$lang['activity_module']                = 'Модуль';
+$lang['activity_modules']               = 'Модули';
+$lang['activity_date']                  = 'Дата';
+$lang['activity_activity']              = 'Действие';
+$lang['activity_not_found']             = 'Действия не найдены.';
+$lang['activity_when']                  = 'Выполено';
+$lang['activity_all_users']             = 'Все пользователи';
+$lang['activity_all_modules']           = 'Все модули';
+$lang['activity_core']                  = 'Ядро Bonfire';
+$lang['activity_all_dates']             = 'В любое время';
+$lang['activity_filter']                = 'Фильтр';
+$lang['activity_filter_head']           = 'Фильтр действий.';
+$lang['activity_filter_note']           = 'Показывать действия %s выбранные %s.';
+$lang['activity_view']                  = 'Отображены %s: <b>%s</b>';
+$lang['activity_delete_own_confirm']    = 'Вы уверены, что хотите удалить ваш лог действий?';
+$lang['activity_delete_user_confirm']   = 'Вы уверены, что хотите удалить лог действий данного пользователя?';
+$lang['activity_delete_module_confirm'] = 'Вы уверены, что хотите удалить лог действий данного модуля?';
+$lang['activity_delete_date_confirm']   = 'Вы уверены, что хотите удалить лог действий за эту дату?';
+$lang['activity_own_delete']            = 'Удалить свои логи';
+$lang['activity_user_delete']           = 'Удалить логи пользователей';
+$lang['activity_module_delete']         = 'Удалить логи модулей';
+$lang['activity_date_delete']           = 'Удалить логи за эту дату';
+$lang['activity_delete_own_note']       = 'Удалить лог активности пользователя:';
+$lang['activity_delete_user_note']      = 'Удалить лог активности пользователя:';
+$lang['activity_delete_module_note']    = 'Удалить лог активности модуля:';
+$lang['activity_delete_date_note']      = 'Удалить лог активности до даты:';
+$lang['activity_own_description']       = 'Ваши действия.';
+$lang['activity_users_description']     = 'Действия пользователей.';
+$lang['activity_module_description']    = 'Действия модулей.';
+$lang['activity_date_description']      = 'Действия по дате.';
+$lang['activity_delete_record']         = 'Удалить лог действий.';
+$lang['activity_delete_confirm']        = 'Вы уверены?';
+$lang['activity_delete_warning']        = 'Данное действие удалит логи активности ниже.';
+$lang['activity_restricted']            = 'Необходима авторизация. Извините, но вы не имеете нужных прав.';
+$lang['activity_nothing']               = 'Найдено 0 записей. Нечего удалять.';
+$lang['activity_deleted']               = '%d запись(ей) удалено.';
+$lang['activity_top_modules']           = '5 самых активных модулей';
+$lang['activity_top_users']             = '5 самых активных пользователей';
+$lang['activity_no_top_modules']        = 'Лог активности модулей пуст.';
+$lang['activity_no_top_users']          = 'Лог активности пользователей пуст.';
+$lang['activity_logged']                = 'Действий';
+$lang['activity_none_found']            = 'Вы не имеете прав для удаления любых действий.';
+$lang['activity_all']                   = 'Все';
+$lang['activity_cleanup']               = 'Очистка логов.';
+
+/* End of file /activities/language/russian/activities_lang.php */