+++ bonfire/modules/docs/language/russian/docs_lang.php
@@ -1,0 +1,32 @@
+<?php defined('BASEPATH') || exit('No direct script access allowed');
+/**
+ * Bonfire
+ *
+ * An open source project to allow developers to jumpstart their development of
+ * CodeIgniter applications
+ *
+ * @package   Bonfire
+ * @author    Bonfire Dev Team
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
+ * @link      http://cibonfire.com
+ * @since     Version 1.0
+ */
+
+/**
+ * Docs language file (Russian).
+ *
+ * Localization strings used by Bonfire.
+ *
+ * @package    Bonfire\Modules\Docs\Language\Russian
+ * @author     Translator < https://github.com/cjmaxik >
+ * @link       http://cibonfire.com/docs/guides
+ */
+
+$lang['docs_title_application'] = 'Приложение';
+$lang['docs_title_bonfire'] = 'Bonfire';
+$lang['docs_title_modules'] = 'Модули';
+
+$lang['docs_not_found'] = 'Искомый документ не найден.';
+
+/* end of file /docs/language/russian/docs_lang.php */