+++ bonfire/modules/migrations/language/russian/migrations_lang.php
@@ -1,0 +1,57 @@
+<?php defined('BASEPATH') || exit('No direct script access allowed');
+/**
+ * Bonfire
+ *
+ * An open source project to allow developers to jumpstart their development of
+ * CodeIgniter applications
+ *
+ * @package   Bonfire
+ * @author    Bonfire Dev Team
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
+ * @link      http://cibonfire.com
+ * @since     Version 1.0
+ */
+
+/**
+ * Migrations language file (Russian).
+ *
+ * Localization strings used by Bonfire.
+ *
+ * @package    Bonfire\Modules\Migrations\Language\Russian
+ * @author     Translator < https://github.com/cjmaxik >
+ * @link       http://cibonfire.com/docs/guides
+ */
+
+// $lang['mig_intro']             = 'Migrations help you keep your database up to date and synced between development and production servers by providing a simple way to keep &lsquo;version control&rsquo; on your database.';
+$lang['mig_not_enabled']       = 'Миграции не включены.';
+$lang['mig_installed_version'] = 'Установленная версия: <b>%d</b>';
+$lang['mig_latest_version']    = 'Последняя доступная версия: <b>%d</b>';
+$lang['mig_db_not_current']    = 'База данных не самая новая.';
+$lang['mig_no_migrations']     = 'Нет доступных миграций.';
+
+// $lang['mig_migrate_note']     = 'Performing migrations <b>WILL</b> change your database structure, possibly ending in disaster. If you are not comfortable with your migrations, please verify them before continuing.';
+$lang['mig_migrate_to']       = 'Мигрировать базу данных на версию';
+$lang['mig_choose_migration'] = 'Мигрировать на версию:';
+$lang['mig_migrate_button']   = 'Миграция базы данных';
+$lang['mig_migrate_module']   = 'Миграция модуля';
+
+$lang['mig_app_migrations']  = 'Миграция приложения';
+$lang['mig_core_migrations'] = 'Миграция ядра Bonfire';
+$lang['mig_mod_migrations']  = 'Миграция модуля';
+
+$lang['db_maintenance'] = 'Обслуживание';
+$lang['db_backups']     = 'Резервные копии';
+$lang['db_migrations']  = 'Миграции';
+
+$lang['mig_tbl_module']        = 'Модулт';
+$lang['mig_tbl_installed_ver'] = 'Установленная версия';
+$lang['mig_tbl_latest_ver']    = 'Последняя версия';
+
+$lang['mig_uninstall'] = 'Ужалить';
+
+$lang['mig_tab_app']  = 'Приложение';
+$lang['mig_tab_mod']  = 'Модули';
+$lang['mig_tab_core'] = 'Bonfire';
+
+/* end of file /migrations/language/russian/migrations_lang.php */