+++ bonfire/modules/permissions/language/russian/permissions_lang.php
@@ -1,0 +1,58 @@
+<?php defined('BASEPATH') || exit('No direct script access allowed');
+/**
+ * Bonfire
+ *
+ * An open source project to allow developers to jumpstart their development of
+ * CodeIgniter applications
+ *
+ * @package   Bonfire
+ * @author    Bonfire Dev Team
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
+ * @link      http://cibonfire.com
+ * @since     Version 1.0
+ */
+
+/**
+ * Permissions language file (Russian).
+ *
+ * Localization strings used by Bonfire.
+ *
+ * @package    Bonfire\Modules\Permissions\Language\Russian
+ * @author     Translator < https://github.com/cjmaxik >
+ * @link       http://cibonfire.com/docs/guides
+ */
+
+$lang['permissions_manage']            = 'Работа с правами';
+$lang['permissions_edit_text']         = 'Измените в соответствии с вашими потребностями.';
+$lang['permissions_no_records']        = 'В системе нет ни одного права';
+$lang['permissions_create_new']        = 'Создать новые права';
+$lang['permissions_create_message']    = 'Создать в системе новые права, которые дадут вам доступ к ним в Ролях.';
+$lang['permissions_create_success']    = 'Права успешно созданы';
+$lang['permissions_create_failure']    = 'Во время создания прав возникла проблема:';
+$lang['permissions_create_new_button'] = 'Создать новые права';
+$lang['permissions_invalid_id']        = 'Неверный ID права';
+$lang['permissions_edit_success']      = 'Права успешно удалены';
+$lang['permissions_edit_failure']      = 'Во время сохранения прав возникла проблема:';
+$lang['permissions_delete_success']    = 'Права были успешно удалены';
+$lang['permissions_delete_failure']    = 'Невозможно удалить права:';
+$lang['permissions_edit_heading']      = 'Изменить права';
+$lang['permissions_save']              = 'Сохранить права';
+$lang['permissions_delete_record']     = 'Удалить эти права';
+$lang['permissions_delete_confirm']    = 'Вы уверены, что хотите удалить эти права?';
+$lang['permissions_delete_warning']    = 'Удаление этих прав также удалит их из ролей.';
+$lang['permissions_id']                = 'ID';
+$lang['permissions_name']              = 'Имя';
+$lang['permissions_description']       = 'Описание';
+$lang['permissions_status']            = 'Статус';
+$lang['permissions_active']            = 'активно';
+$lang['permissions_inactive']          = 'неактивно';
+$lang['permissions_permission']        = 'Права';
+$lang['permissions_deleted']           = 'Права удалены';
+$lang['permissions_del_failure']       = 'Невозможно удалить права:';
+$lang['permissions_del_error']         = 'Не выбраны права для удаления.';
+$lang['permissions_intro']             = 'Права обеспечивают точный контроль над тем, что каждая роль может делать.';
+$lang['permissions_matrix']            = 'Матрица прав';
+$lang['permissions_details']           = 'Детали прав';
+
+/* end of file /permissions/language/russian/permissions_lang.php */