+++ bonfire/modules/sysinfo/language/russian/sysinfo_lang.php
@@ -1,0 +1,39 @@
+<?php defined('BASEPATH') || exit('No direct script access allowed');
+/**
+ * Bonfire
+ *
+ * An open source project to allow developers to jumpstart their development of
+ * CodeIgniter applications
+ *
+ * @package   Bonfire
+ * @author    Bonfire Dev Team
+ * @copyright Copyright (c) 2011 - 2014, Bonfire Dev Team
+ * @license   http://opensource.org/licenses/MIT
+ * @link      http://cibonfire.com
+ * @since     Version 1.0
+ */
+
+/**
+ * Sysinfo language file (Russian).
+ *
+ * Localization strings used by Bonfire.
+ *
+ * @package    Bonfire\Modules\Sysinfo\Language\Russian
+ * @author     Translator < https://github.com/cjmaxik >
+ * @link       http://cibonfire.com/docs/guides
+ */
+
+$lang['si_system_info'] = 'Системная информация';
+$lang['si_system']      = 'Система';
+$lang['si_modules']     = 'Модули';
+$lang['si_php']         = 'PHP';
+
+$lang['si_installed_mods'] = 'Установленные модули';
+$lang['si_php_info']       = 'Информация PHP';
+
+$lang['sys_mod_name']   = 'Имя модуля';
+$lang['sys_mod_ver']    = 'Версия';
+$lang['sys_mod_desc']   = 'Описание';
+$lang['sys_mod_author'] = 'Автор';
+
+/* /sysinfo/language/russian/sysinfo_lang.php */